Workaround QML renderer that appears stuck
After wakeup of laptop it seems that QML is not responding anymore.
Actions like menu items, selecting UI items, selecting dropdown items
can be triggered but there's no visual rendering in the QML area
happening anymore.
It seems that opening the about dialog or resizing the main window
is a workaround of the issue !?